fix toString
'0' isn't in the charset of base58.  Instead, '1' is the character with index 0 in the base58 charset.  Also, instead of rounding to the appropriate bits, just return nothing if it is out of range.
Also, I guess my impl is little endian, not big endian so it's padEnd instead of padStart.