Firefox, please.
So, this is still only like 50-50.  Which probably means that this only works if the ICE role conflict resolution mechanism resolves to either a or b.  But I don't know which.  And I need to make it work no matter which is the winner of the role.

Technically, I need to just switch back to writing my own TURN server and hopefully these FF changes will have improved that situation somewhat.