pain, but finally decrypting
yay!  We're down to the last protocol: SCTP.

Once we've implemented SCTP then we can send data from the server to the client which is how we'll implement a bind server of sorts.  The server will notify browser peers that there is someone trying to connect to them by passing the STUN binding request packet as a binary data message on some kind of bin datachannel.