DTLS client or server!
The relay/switch can now be either the DTLS client or server meaning is follows the same rules as browser<->browser connections.
There's still big problems though.  The shared controller means things won't shutdown cleanly in some cases.  It's a complete abuse of the transform stream stuff, but that can be fixed later.