change default setup
Previously, I had the DTLS role (client / server) determined via politeness, but I've decided that I would rather have the DTLS role determined via who is using the address.  We can still access the previous functionality (setup role determined via politeness) by setting setup=<empty string> search parameter because then the Conn wrapper will assign the setup roles using politeness (as before).

This change partially simplifies hosted peers because they now only need to be a DTLS server (before they needed to sometimes be a client).

I've been worried about denial of service surface area and having the DTLS roles always determined by peer id seemed like a potential problem.

There was one reason I held off on this and that is because it may be beneficial for the owner of an address to actually be the client instead of the server.  This could be the case if we use two RTCPeerConnections: one to gather candidate and a second one to answer incoming connections.  In this scenario it we don't wan the calling peer to attempt the DTLS handshake with our gathering RTCPC because that would cause the connection to fail.  Under this new default, you would use setup=active for this case.