separate direct vs indirect
We only provide the xpeer when forwarding a connection test on to the peer who is the intended recipient of that connection test and therefore could answer it by turning it + the xpeer into an ICE candidate.

When forwarding to a random person we only provide the contents of the packet which is enough for the receiver to check the password used and to know the lufrag and rufrag.

You wouldn't be able to build and ICE candidate from this because you have no ip+port, but you might decide to initiate and outgoing connection to either lufrag or rufrag.  I've added this functionality to potentially support bootstrapping networks in the future using a custom password.