relay
This is roughly the same relay server as masquerade except masquerade has an inner hosted service.  I've been having trouble with that service (which was fake-ICE and weird DTLS, etc.) so I'm thinking I need an actual webrtc impl to answer incoming connections that we discover.  Not quite sure how I'll be doing that yet.